URLs now have http(s) stripped before being passed to `dns.lookup()`
So now we are able to take any url and determine if it is a valid url,
either by looking at it and determining it has no protocol information
or asking `dns.lookup()` what it thinks.

What i'm working on right now is to ask the database if the full url
is in any of the records. But having issues with how to get the data out
of `Model.find()`.